pushing for latest updates to project

+++ controllers/clothingItem.js
@@ -15,13 +15,38 @@ const createItem = (req, res) => {
   ClothingItem.create({ name, weather, imageURL })
     .then((item) => {
       console.log(item);
+      console.log("im here in then");
       res.send({ data: item });
     })
     .catch((e) => {
       console.log(e.name);
-      res
-        .status(INTERNAL_SERVER_ERROR_CODE)
-        .send({ message: "Error from createItem", e });
+      console.log(name);
+      console.log(weather);
+      console.log(imageURL);
+      console.log("im here in catch");
+      if (e.name === "ValidationError" && name === undefined) {
+        res.status(VALIDATION_ERROR_CODE).send({ message: "Name is invalid" });
+      } else if (e.name === "ValidationError" && weather === undefined) {
+        res
+          .status(VALIDATION_ERROR_CODE)
+          .send({ message: "weather is invalid" });
+      } else if (e.name === "ValidationError" && imageURL === undefined) {
+        res
+          .status(VALIDATION_ERROR_CODE)
+          .send({ message: "imageURL is invalid" });
+      } else if (e.name === "ValidationError" && name.length < 2) {
+        res
+          .status(CAST_ERROR_ERROR_CODE)
+          .send({ message: "Name is under character limit" });
+      } else if (e.name === "ValidationError" && name.length > 30) {
+        res
+          .status(CAST_ERROR_ERROR_CODE)
+          .send({ message: "Name is past character limit" });
+      } else {
+        res
+          .status(INTERNAL_SERVER_ERROR_CODE)
+          .send({ message: "Error from createItem", e });
+      }
     });
 };
 
@@ -64,36 +89,58 @@ const deleteItem = (req, res) => {
 
   console.log(itemId);
   ClothingItem.findByIdAndDelete(itemId)
-    .orFail(() => {
-      const error = new Error("Item ID not found");
-      error.statusCode = 404;
-      throw error;
-    })
+    // .orFail(() => {
+    //   const error = new Error("Item ID not found");
+    //   error.statusCode = 404;
+    //   throw error;
+    // })
     .then((item) => {
-      res.status(204).send({});
+      console.log("im in then");
+      res.status(200).send({ message: "successfully deleted" });
     })
     .catch((e) => {
       console.log(e.name);
-      res
-        .status(INTERNAL_SERVER_ERROR_CODE)
-        .send({ message: "Error from deleteItem", e });
+      console.log(e);
+      console.log("im in catch");
+      if (e.name === "CastError") {
+        res
+          .status(CAST_ERROR_ERROR_CODE)
+          .send({ message: "incorrect or _id or _id does not exist " });
+      } else if (e.name === "Error") {
+        res
+          .status(NOT_FOUND_ERROR_CODE)
+          .send({ message: "_id was not found or does not exist" });
+      } else {
+        res
+          .status(INTERNAL_SERVER_ERROR_CODE)
+          .send({ message: "Error from deleteItem", e });
+      }
     });
 };
 
-module.exports.likeItem = (req, res) =>
+const likeItem = (req, res) =>
   ClothingItem.findByIdAndUpdate(
     req.params.itemId,
     { $addToSet: { likes: req.user._id } },
     { new: true },
-  );
+  )
+    .then((like) => {
+      res.status(200).send({ data: like });
+    })
+    .catch((e) => {
+      console.log("im in catch for likeItem");
+      console.log(e);
+    });
 // ...
 
-module.exports.dislikeItem = (req, res) =>
+const dislikeItem = (req, res) =>
   ClothingItem.findByIdAndUpdate(
     req.params.itemId,
     { $pull: { likes: req.user._id } },
     { new: true },
-  );
+  ).catch((e) => {
+    console.log(e);
+  });
 // ...
 
 module.exports = {
@@ -101,4 +148,6 @@ module.exports = {
   getItems,
   updateItem,
   deleteItem,
+  likeItem,
+  dislikeItem,
 };
